requirements.txt
pandas
scikit-learn
kagglehub
python-dotenv
torch
sentence-transformers
google-generativeai
faiss-cpu